Wi-Fi tethering switch was briefly in wrong state
If Wi-Fi tethering is on when entering the "Tethering &
portable hotspot" activity, the switch for Wi-Fi tethering
is briefly displayed as off and then goes to on. The icon in
status bar indicates that Wi-Fi tethering is on during this
behaviour, so it is only the switch that is not correctly set
when the activity is started. To fix this, the state of Wi-Fi
tethering is read from the WifiManager when the activity is
created.

Change-Id: I3891b84a2d949c1573b3501d2fcba9945d4590a1